requirements.txt
APScheduler
django_apscheduler
cx-Oracle
Django
pytz
elasticsearch
requests
sqlalchemy
mysqlclient
tzlocal
dominate
openpyxl
djangorestframework
djangorestframework-jwt
django-cors-middleware
kubernetes
uuid
rgwadmin
ansible
channels
channels_redis
python-jenkins
DBUtils